+++ src/sql/src/normalize.rs
@@ -237,6 +237,7 @@ pub fn create_statement(scx: &StatementContext, mut stmt: Statement) -> Result<S
             name: _,
             on_name,
             key_parts,
+            with_options: _,
             if_not_exists,
         }) => {
             *on_name = resolve_item(on_name)?;